explicit wheel event handling on address input popover content
Much like the pointer events, mouse wheel events were
also not being recognised. However, the handling for
this is a little more elaborate, because after we create
the explicit handling, we also need to stop additional
porpagation to avoid jerky scroll behavior. Maybe there's
a better way to handle this but it elludes me.